Offer debugger methods to dump a variable
Summary:
Import `dprint_r` and `dieprint_r` to the library.

These functions allow to enclose a print_r call in a preformatted block,
so the output could be read directly in a rendered HTML document.

They are intended to be used as debug time.

As these global functions are still in use during some current debug
and development tasks, we offer a mechanism to register them from the
Debugger class. As `create_function` is deprecated since PHP 7.2.0,
the selected method is to require a wrapper file in the global space.

The result is a choice between those two syntaxes:

```lang=php
Debugger::register();
dprint_r($someVariable);
```

```lang=php
Debugger::printVariable($someVariable);
```

Ref T1395

Test Plan:
  - unit test provided for the registration
  - integration tests provided for the debug methods

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1395

Differential Revision: https://devcentral.nasqueron.org/D1614