Provide a data tests folder with MockLib
Summary:
To move include code from the Autoloader -er class to a CodeFile class,
we need an unit test to check if we've included a file. There are such
files in MockLib folder. That would be more convenient to move this
folder in a data/ folder.

The WithData trait allows to get path from this folder.

Test Plan: Tests still run correctly

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D1618