Allow to get server URL, port, name, scheme
Summary:
To determine the URL of a server, three components are needed:
the scheme (http? https?), the name and the port.

Generally, ports won't be the in URL, as they are the default
80 or 443 ports.

**Advanced network configurations**

Furthermore, in back-end/front-end topologies, To determine
if we're in HTTPS requires to examine several headers.

If there is now a standard, RFC 7239 section 4, a lot of load balancers,
caches or web servers use proprietary or former de facto headers.

**Deprecation**

The new Request::getServerURL method supersedes get_server_url
from Keruald/Xen sites like Azhàr or Zed.

Test Plan: Unit tests provided

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D1636