Address File and CodeFile inheritance issues
Summary:
The File class also implemented include logic.

Actually, this code was the one used when object was instancied by
CodeFile::From(), as it returned a File and not a CodeFile instance.

That first issue masked another one, CodeFile included the file
by name, and not by full path.

Test Plan: Run tests

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D2162