Provide URL
Summary:
**Context**

The core.php or config.php files in Xen and Pluton tend to provide
the ability to build URLs relative to the server.

To help to do so, we provide an URL class implementation with
understanding of protocol, domain and query part.

**Implementation**

IDN are supported.

By default, queries are encoded excepted for / ~ . - _ characters.

**OmniArray and OmniString**

This change introduces OO-variants of map, implode, explode
to handle the "encode everything excepted slashes" task.

Test Plan: Unit tests provided

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D2282