Output report in HTML and Markdown
Summary:
**HTML implementation notes**

HTML is more useful as a fragment than as full page,
to be able to be included in a regular application.

An entry can be divided as several paragraphs,
as it could represent several ideas.

Ref T1678

Test Plan:
Unit tests for a regular report are provided.

Only missing some cases if the rapport, or a part of it like properties, is empty.

Reviewers: dereckson

Reviewed By: dereckson

Subscribers: dereckson

Maniphest Tasks: T1678

Differential Revision: https://devcentral.nasqueron.org/D2505