Allow to setup the reository remotes
Summary:
The CD job to push from this monorepo to the packages' manyrepo
needs to get remotes properly configured.

This script allows to use devcentral hostname and default port 22
so it can be run from a container belonging to the same network
than DevCentral (or with bridged access to).

Ref T1680.

Test Plan:

```
$ git remote remove report
$ _utils/repo/setup.sh
$ git remote
```

Reviewers: dereckson

Differential Revision: https://devcentral.nasqueron.org/D2543