Support blackhole database engine for mocks
Summary:
It can be convenient in tests to replace a full fledged instance
of the database engine by a mock doing nothing. That's the purpose
of this change.

This blackhole engine doesn't really use any of the parameters passed
to the methods, and that's expected behavior for a blackhole. As such,
phpcs is configured to ignore unused parameters.

Test Plan: used in Zed to test GeoLocation class

Reviewers: dereckson

Differential Revision: https://devcentral.nasqueron.org/D3198